fix(resolv): fix the bug that the parseKeyword func of the querier parses the keyword value abnormally
Use the keyword of `<ParserType>:sep: <key and value string>` type. When the value length of `<key and value string>` is less than 5, an error will be reported in judging whether the keyword is a regular matching rule because the value length of `<key and value string>` is less than 5.